remove redundant interface
due to it breaking implicit casts with static checks
