+++ Directory.Build.props
@@ -12,7 +12,7 @@
     <NoWarn>$NoWarn;1591;CA1014;CA1062;CA2007;CA1822</NoWarn>
   </PropertyGroup>
   <PropertyGroup>
-    <VersionPrefix>4.3.1</VersionPrefix>
+    <VersionPrefix>5.0.0</VersionPrefix>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.2.32" PrivateAssets="all">
