+++ Mindbox.I18n.Abstractions/LocalizableString.cs
@@ -1,29 +1,26 @@
 // Copyright 2022 Mindbox Ltd
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using Mindbox.I18n.Abstractions;
+namespace Mindbox.I18n.Abstractions;
 
-namespace Mindbox.I18n;
-
-public abstract class LocalizableString : ILocalizableString
+public abstract class LocalizableString
 {
-	private static ILogger Logger { get; set; } = new NullI18NextLogger();
+	private static ILogger? Logger { get; set; }
 
 	public static void InitializeLogger(ILogger logger)
 	{
-		Logger = logger ?? new NullI18NextLogger();
+		Logger = logger;
 	}
 
 	public static LocalizableString ForKey([LocalizationKey] string key)
@@ -51,7 +48,7 @@ public abstract class LocalizableString : ILocalizableString
 
 	public override string ToString()
 	{
-		Logger.LogInvalidOperation($"ToString() called on LocalizableString with key {Key}");
+		Logger?.LogInvalidOperation($"ToString() called on LocalizableString with key {Key}");
 		return Key;
 	}
