+++ Mindbox.I18n.Abstractions/LocalizationKeyAttribute.cs
@@ -1,20 +1,21 @@
 // Copyright 2022 Mindbox Ltd
-// 
+//
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
-// 
+//
 //     http://www.apache.org/licenses/LICENSE-2.0
-// 
+//
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+
 namespace Mindbox.I18n.Abstractions;
 
-public interface ILocalizableString
+[AttributeUsage(validOn: AttributeTargets.Field | AttributeTargets.Property | AttributeTargets.Parameter)]
+public sealed class LocalizationKeyAttribute : Attribute
 {
-	public abstract string Render(ILocalizationProvider localizationProvider, ILocale locale);
 }
